+++ src/components/ReviewCard/Action.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { PiHeartFill } from "react-icons/pi";
 
-import { Reaction, LikeButton } from "../style";
+import { Reaction, LikeButton } from "./style";
 
 export default function ReviewAction() {
   const [isLike, setIsLike] = useState(false);
